PUT Candidato y horas
En el put de candidatos ya se puede votar y modificar el candidato (alguna de las dos o las dos a la vez)
{"votar":{"idUsuario":idDelUsuario,"voto":0}}
o
{"candidato":{cosas del candidato}}
o las dos a la vez
{"votar":{"idUsuario":idDelUsuario,"voto":0},"candidato":{cosas del candidato}}

La hora era cambiar getDate por getTime en los mappers

+++ src/java/rest/CandidatosResource.java
@@ -29,6 +29,8 @@ import modelo.Candidatos;
 import modelo.ErrorSimple;
 import modelo.ErrorYID;
 import modelo.IDUsuario;
+import modelo.IDUsuarioYVoto;
+import modelo.IDUsuarioYVotoYCandidato;
 import modelo.Usuario;
 import modelo.UsuarioVotaCandidato;
 import modelo.VotoIDUsuario;
@@ -88,6 +90,7 @@ public class CandidatosResource {
     public ErrorYID newCandidato(Candidato candidato){
         candidato.setId(0);
         candidato.setFecha(new Date());
+        candidato.setEstado(0);
         int nuevoID;
         if(candidato.getGestionadoPor() <= 0){
             nuevoID = candidatoMapperSinGestionadoPor.insert(candidato);
@@ -110,35 +113,53 @@ public class CandidatosResource {
     }
     
     @PUT
-    @Path("{idCandidato}/{voto}")
+    @Path("{idCandidato}")
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
     //Poner en raw el id del usuario: {"idUsuario":1}
     //poner en la url /candidatos/3/1  (3 es el idCandidato y 1 es voto positivo, 0 es negativo)
-    public ErrorSimple putCandidato(@PathParam("idCandidato") int idCandidato, IDUsuario idUsuario, @PathParam("voto") int posNeg) {
-        Candidato aModificar = candidatoMapper.findById(idCandidato);
-        Usuario usuario = usuarioMapper.findById(idUsuario.getIdUsuario());
-            String error = "si";
+    public ErrorSimple putCandidato(@PathParam("idCandidato") int idCandidato, IDUsuarioYVotoYCandidato iDUsuarioYVotoYCandidato) {
+        IDUsuarioYVoto idUsuario = iDUsuarioYVotoYCandidato.getiDUsuarioYVoto(); 
+        Candidato candidato = iDUsuarioYVotoYCandidato.getCandidato();
+        String error = "si";
+        
+        if (idUsuario != null){
+        int posNeg = idUsuario.getVoto();
+            Candidato aModificar = candidatoMapper.findById(idCandidato);
+            Usuario usuario = usuarioMapper.findById(idUsuario.getIdUsuario());
+                
 
-            if(aModificar != null && usuario != null){
-                UsuarioVotaCandidato nuevo = new UsuarioVotaCandidato(idUsuario.getIdUsuario(), idCandidato, posNeg);
+                if(aModificar != null && usuario != null){
+                    UsuarioVotaCandidato nuevo = new UsuarioVotaCandidato(idUsuario.getIdUsuario(), idCandidato, posNeg);
+                    error = "no";
+                    //Si no existe lo creamos, si existe lo borramos
+                    UsuarioVotaCandidato antiguo = usuarioVotaCandidatoMapper.findById(nuevo); //Si el nuevo ya existe se guarda en antiguo
+                    if (antiguo == null){
+                        usuarioVotaCandidatoMapper.insert(nuevo);
+                    }
+                    else {
+                        if(antiguo.getVoto() == nuevo.getVoto())
+                            usuarioVotaCandidatoMapper.delete(antiguo);
+                        else usuarioVotaCandidatoMapper.update(nuevo);
+                    }
+                }
+
+        }
+        
+        if (candidato != null){
+            Candidato aModificar = candidatoMapper.findById(idCandidato);
+            if (aModificar != null){
+                candidato.setId(idCandidato);
+                candidato.setFecha(aModificar.getFecha());
+                candidatoMapper.update(candidato);
                 error = "no";
-                //Si no existe lo creamos, si existe lo borramos
-                UsuarioVotaCandidato antiguo = usuarioVotaCandidatoMapper.findById(nuevo); //Si el nuevo ya existe se guarda en antiguo
-                if (antiguo == null){
-                    usuarioVotaCandidatoMapper.insert(nuevo);
-                }
-                else {
-                    if(antiguo.getVoto() == nuevo.getVoto())
-                        usuarioVotaCandidatoMapper.delete(antiguo);
-                    else usuarioVotaCandidatoMapper.update(nuevo);
-                }
             }
+        }
         
         return new ErrorSimple(error);
     } 
     
-    
+    /**
     //Para que un usuario administrador (rol != 0) pueda gestionar un candidato
     //se usa el Put del candidato con el id del usuario administrador en raw.
     @PUT
@@ -161,7 +182,7 @@ public class CandidatosResource {
         }
         return new ErrorSimple(error);
     }
-    
+    **/
     
     @DELETE
     @Path("{idCandidato}")
